Time: 36 ms (14.42%), Space: 22.4 MB (72.97%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -30,7 +30,7 @@ public:
                 levelSum += child->val;
             }
             
-            res.push_back(levelSum/n);
+            res.push_back((double)levelSum/n);
             
         }
         return res;
